Crawler/requirements.txt
selenium
pymysql
beautifulsoup4
python-dotenv
requests